handle error for export also
see 1b2159f for the initial implementation on handling the generation error.
this change now handles also the case for font export.

additionally, we first generate the export by ajax and only after redirect to
the download. by adding caching the 2nd call to the same export should not
be a huge overhead.