update js libs (fd-slider)
i had to add the fd slider stock css file, which is huge
and actually the doc says it should be generated. the generatoin
process is pretty weird though (includes mhtml/png files and
absolute server urls), so i left it as is.
some day maybe i should extract only the necessary parts.